Refactor GhostBox and Element
